make AcountTypes constants uint8

+++ types/types.go
@@ -12,7 +12,7 @@ import (
 
 // AccountType is the type of account. It should be same as the type defined in the crates/types/src/account.rs and libinitia/src/api.rs
 const (
-	AccountType_Base = 0 + iota
+	AccountType_Base = uint8(0) + iota
 	AccountType_Object
 	AccountType_Table
 	AccountType_Module
